Renaming libraries as per naming conventions
