+++ drivers/unifiedPlatform/platform/backend/v1/api/tenant.go
@@ -1,4 +1,4 @@
-package apiv1
+package api
 
 import (
 	"context"
