+++ drivers/unifiedPlatform/unifiedPlatform.go
@@ -3,6 +3,8 @@ package unifiedPlatform
 import (
 	"crypto/tls"
 	"fmt"
+	. "github.com/portworx/torpedo/drivers/unifiedPlatform/pds/backend/v1/api"
+	pdsGrpc "github.com/portworx/torpedo/drivers/unifiedPlatform/pds/backend/v1/grpc"
 	"github.com/portworx/torpedo/pkg/log"
 	"os"
 	"strconv"
@@ -10,11 +12,9 @@ import (
 
 	pdsv2 "github.com/portworx/pds-api-go-client/unifiedcp/v1alpha1"
 	"github.com/portworx/torpedo/drivers/unifiedPlatform/pds"
-	. "github.com/portworx/torpedo/drivers/unifiedPlatform/pds/backend/apiv1"
-	pdsGrpc "github.com/portworx/torpedo/drivers/unifiedPlatform/pds/backend/grpc"
 	"github.com/portworx/torpedo/drivers/unifiedPlatform/platform"
-	. "github.com/portworx/torpedo/drivers/unifiedPlatform/platform/backend/apiv1"
-	platformGrpc "github.com/portworx/torpedo/drivers/unifiedPlatform/platform/backend/grpc"
+	. "github.com/portworx/torpedo/drivers/unifiedPlatform/platform/backend/v1/api"
+	platformGrpc "github.com/portworx/torpedo/drivers/unifiedPlatform/platform/backend/v1/grpc"
 	. "github.com/portworx/torpedo/drivers/utilities"
 	platformv1 "github.com/pure-px/platform-api-go-client/v1alpha1"
 	"google.golang.org/grpc"
